feat: move actualites component
