+++ src/app/components/dashboards/actualites/actualites.component.html
@@ -1,0 +1,1 @@
+<p>actualites works!</p>
