+++ src/app/components/dashboards/sidebar/sidebar.component.html
@@ -14,5 +14,8 @@
   <div class="onglet">
     <a routerLink="evenements">Evénements</a>
   </div>
+  <div class="onglet">
+    <a routerLink="actualites">Actualités</a>
+  </div>
 </div>
