--- src/app/pages/actualites/actualites.component.html
@@ -1,2 +1,0 @@
-<p>actualites works!</p>
-<app-footer></app-footer>
