+++ src/app/pages/dashboard/dashboard.component.ts
@@ -1,3 +1,4 @@
+import { UserService } from './../../shared/services/user.service';
 import { DashboardService } from './../../shared/services/dashboard.service';
 import { Component, OnInit } from '@angular/core';
 import { User } from 'src/app/shared/models/user';
@@ -8,8 +9,9 @@ import { User } from 'src/app/shared/models/user';
   styleUrls: ['./dashboard.component.scss']
 })
 export class DashboardComponent implements OnInit {
+  UserService: any;
 
-  constructor(private service: DashboardService) { }
+  constructor(private service: DashboardService, private userService: UserService) { }
 
 
   user: User;
@@ -17,11 +19,7 @@ export class DashboardComponent implements OnInit {
 
   ngOnInit() {
 
-  }
-  changeStatus() {
-    this.StatusAccountIsComplete = !this.StatusAccountIsComplete;
-  }
-  getAccountStatus() {
+    this.user = this.userService.user1;
 
   }
 }
